+++ app/assets/stylesheets/main.scss
@@ -13,11 +13,11 @@ body {
 }
 
 .flash {
+  text-align: center;
   background: $background;
   color: $highlight;
-  padding-left: 1rem;
-  padding-top: 1rem;
-  font-family: fantasy;
+  padding: 1rem 0.75rem;
+  font-family: 'Josefin';
   font-weight: bold;
   text-shadow: 1px 1px #000000;
 }
@@ -34,7 +34,7 @@ body {
 }
 
 #items-index-container {
-  padding:0;
+  padding: 0;
 }
 
 #item-show-component {
@@ -44,8 +44,6 @@ body {
   position: relative;
 
   .item-show-image {
-    // margin: 0.1rem;
-    // border-style: solid;
     padding: 1.5rem;
     display:inline-block;
     width: 450px;
@@ -132,7 +130,6 @@ body {
   #button:hover {
     box-shadow: 0 12px 16px 0 rgba(0,0,0,0.24), 0 17px 50px 0 rgba(0,0,0,0.19);
     font-weight: bold;
-
     font-family: 'Josefin';
     color: white;
     padding: 0.75rem;
@@ -159,7 +156,6 @@ body {
 
     li:nth-child(n)::before {
       content: "❌ ";
-
     }
   }
 }
@@ -239,7 +235,6 @@ body {
       width: 100px;
       box-shadow: 0 12px 16px 0 rgba(0,0,0,0.24), 0 17px 50px 0 rgba(0,0,0,0.19);
       font-weight: bold;
-
       font-family: 'Josefin';
       color: white;
       padding: 0.75rem;
@@ -274,7 +269,6 @@ body {
 
     #tile-field {
       padding-top: 1rem;
-
     }
   }
 }
@@ -359,6 +353,7 @@ body {
 }
 
 #item-tile {
+
   #button {
     font-family: 'Josefin';
     color: white;
@@ -372,7 +367,6 @@ body {
   #button:hover {
     box-shadow: 0 12px 16px 0 rgba(0,0,0,0.24), 0 17px 50px 0 rgba(0,0,0,0.19);
     font-weight: bold;
-
     font-family: 'Josefin';
     color: white;
     padding: 0.75rem;
@@ -386,6 +380,7 @@ body {
 #item-search-component {
 
   #item-search-form {
+
     #search {
       margin: 0;
     }
@@ -407,7 +402,7 @@ body {
       padding: 0.75rem;
       margin: 0;
       width: 100%;
-      background: #E8175D;
+      background: #EB0A44;
       border: none;
     }
   }
@@ -422,6 +417,7 @@ body {
     padding: 1rem;
     margin: 0;
   }
+
   #button:hover {
     color: white;
     font-weight: bold;
