+++ config/routes.rb
@@ -3,6 +3,10 @@ Rails.application.routes.draw do
   devise_for :users
   # For details on the DSL available within this file, see http://guides.rubyonrails.org/routing.html
 
+  unauthenticated :user do
+    root to: 'devise/sessions#new'
+  end
+
   get '/items', to: 'homes#index'
   get '/items/:id', to: 'homes#index'
   get '/users/:id', to: 'homes#index'
